Codecov and simply code (#2159)
* code-coverage: improve zlib compression and decompression error handling

* refactor: improve logging format for better readability

* refactor: simplify switch statements for property rotation

* refactor: ensure proper long type casting in various methods

+++ src/main/java/cn/nukkit/utils/Utils.java
@@ -238,7 +238,7 @@ public class Utils {
 
     public static long toABGR(int argb) {
         long result = argb & 0xFF00FF00L;
-        result |= (argb << 16) & 0x00FF0000L; // B to R
+        result |= ((long) argb << 16) & 0x00FF0000L; // B to R
         result |= (argb >>> 16) & 0xFFL; // R to B
         return result & 0xFFFFFFFFL;
     }
